refactor: move llm folder in lib/api
