+++ front/lib/api/llm/types/options.ts
@@ -1,0 +1,6 @@
+import type { AgentReasoningEffort } from "@app/types";
+
+export type LLMOptions = {
+  reasoningEffort?: AgentReasoningEffort;
+  temperature?: number;
+};
